[Feature] 完成岗位用户、角色权限关联

+++ ace-admin/src/main/java/com/github/wxiaoqi/security/admin/biz/DepartBiz.java
@@ -5,11 +5,14 @@ import com.github.wxiaoqi.merge.annonation.MergeResult;
 import com.github.wxiaoqi.security.admin.entity.Depart;
 import com.github.wxiaoqi.security.admin.entity.User;
 import com.github.wxiaoqi.security.admin.mapper.DepartMapper;
+import com.github.wxiaoqi.security.admin.mapper.UserMapper;
 import com.github.wxiaoqi.security.admin.service.TableResultParser;
 import com.github.wxiaoqi.security.common.biz.BusinessBiz;
+import com.github.wxiaoqi.security.common.exception.base.BusinessException;
 import com.github.wxiaoqi.security.common.msg.TableResultResponse;
 import com.github.wxiaoqi.security.common.util.UUIDUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
@@ -26,6 +29,8 @@ import java.util.stream.Collectors;
  */
 @Service
 public class DepartBiz extends BusinessBiz<DepartMapper,Depart> {
+    @Autowired
+    private UserMapper userMapper;
     @MergeResult(resultParser = TableResultParser.class)
     public TableResultResponse<User> getDepartUsers(String departId,String userName) {
         List<User> users = this.mapper.selectDepartUsers(departId,userName);
@@ -56,6 +61,10 @@ public class DepartBiz extends BusinessBiz<DepartMapper,Depart> {
     }
 
     public void delDepartUser(String departId, String userId) {
+        User user = userMapper.selectByPrimaryKey(userId);
+        if(user.getDepartId().equals(departId)){
+            throw new BusinessException("无法移除用户的默认关联部门,若需移除,请前往用户模块更新用户部门!");
+        }
         this.mapper.deleteDepartUser(departId,userId);
     }
 }
